+++ Rdmp.UI/PieCharts/GoodBadCataloguePieChart.cs
@@ -5,22 +5,21 @@
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
 using System;
+using System.Collections.Generic;
 using System.Data;
 using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
 using Rdmp.Core.Curation.Data;
 using Rdmp.Core.Curation.Data.Dashboarding;
+using Rdmp.UI.CommandExecution.AtomicCommands;
 using Rdmp.UI.DashboardTabs.Construction;
 using Rdmp.UI.Icons.IconProvision;
 using Rdmp.UI.ItemActivation;
 using Rdmp.UI.Refreshing;
 using Rdmp.UI.SimpleDialogs;
 using Rdmp.UI.TestsAndSetup.ServicePropogation;
-using ReusableLibraryCode.Icons.IconProvision;
-using ReusableUIComponents;
 using ReusableUIComponents.Dialogs;
-using ReusableUIComponents.SingleControlForms;
 
 namespace Rdmp.UI.PieCharts
 {
@@ -37,7 +36,9 @@ namespace Rdmp.UI.PieCharts
         private ToolStripButton btnRefresh = new ToolStripButton("Refresh",FamFamFamIcons.text_list_bullets) { Name = "btnRefresh" };
         private ToolStripLabel toolStripLabel1 = new ToolStripLabel("Type:"){Name= "toolStripLabel1" };
         private ToolStripButton btnShowLabels = new ToolStripButton("Labels",FamFamFamIcons.text_align_left) { Name = "btnShowLabels", CheckOnClick = true };
-        
+
+        private List<ToolStripButton> _flagOptions = new List<ToolStripButton>();
+
         public GoodBadCataloguePieChart()
         {
             InitializeComponent();
@@ -52,7 +53,33 @@ namespace Rdmp.UI.PieCharts
             //put edit mode on for the designer
             NotifyEditModeChange(false);
         }
-        
+
+        private void SetupFlags()
+        {
+            if(!firstTime)
+                return;
+
+            AddFlag("Non Extractable Catalogues", c=> c.IncludeNonExtractableCatalogues, (c,r) => c.IncludeNonExtractableCatalogues= r);
+            AddFlag("Deprecated Catalogues", c => c.IncludeDeprecatedCatalogues, (c, r) => c.IncludeDeprecatedCatalogues = r);
+            AddFlag("Internal Catalogues", c => c.IncludeInternalCatalogues, (c, r) => c.IncludeInternalCatalogues = r);
+            AddFlag("Cold Storage Catalogues", c => c.IncludeColdStorageCatalogues, (c, r) => c.IncludeColdStorageCatalogues = r);
+            AddFlag("Project Specific Catalogues", c => c.IncludeProjectSpecificCatalogues, (c, r) => c.IncludeProjectSpecificCatalogues = r);
+
+            AddFlag("Non Extractable CatalogueItems", c => c.IncludeNonExtractableCatalogueItems, (c, r) => c.IncludeNonExtractableCatalogueItems = r);
+            AddFlag("Internal Catalogue Items", c => c.IncludeInternalCatalogueItems, (c, r) => c.IncludeInternalCatalogueItems = r);
+            AddFlag("Deprecated Catalogue Items", c => c.IncludeDeprecatedCatalogueItems, (c, r) => c.IncludeDeprecatedCatalogueItems = r);
+            firstTime = false;
+        }
+
+        private void AddFlag(string caption, Func<GoodBadCataloguePieChartObjectCollection,bool> getProp, Action<GoodBadCataloguePieChartObjectCollection,bool> setProp)
+        {
+            var btn = new ToolStripButton(caption);
+            btn.Checked = getProp(_collection);
+            btn.CheckedChanged += (sender,e) =>{setProp(_collection,((ToolStripButton)sender).Checked);};
+            btn.CheckedChanged += (s, e) => GenerateChart();
+            btn.CheckOnClick = true;
+            _flagOptions.Add(btn);
+        }
         private DashboardControl _dashboardControlDatabaseRecord;
         private GoodBadCataloguePieChartObjectCollection _collection;
         
@@ -73,23 +100,9 @@ namespace Rdmp.UI.PieCharts
         {
             try
             {
-                ExtractionInformation[] allExtractionInformation;
+                CatalogueItem[] catalogueItems = GetCatalogueItems();
 
-                if (!_collection.IsSingleCatalogueMode)
-                {
-                    //get the active (non depricated etc) Catalogues
-                    var activeCatalogues = Activator.CoreChildProvider.AllCatalogues.Where(ShouldHaveDescription).ToArray();
-                        
-                    //if there are some
-                    if(activeCatalogues.Any())
-                        allExtractionInformation = activeCatalogues.SelectMany(c=>c.GetAllExtractionInformation(ExtractionCategory.Any)).ToArray();//get the extractable columns
-                    else
-                        allExtractionInformation = new ExtractionInformation[0];//there weren't any so Catalogues so wont be any ExtractionInformationsEither
-                }
-                else
-                    allExtractionInformation = _collection.GetSingleCatalogueModeCatalogue().GetAllExtractionInformation(ExtractionCategory.Any);
-
-                if (!allExtractionInformation.Any())
+                if (!catalogueItems.Any())
                 {
                     chart1.DataSource = null;
                     chart1.Visible = false;
@@ -101,8 +114,8 @@ namespace Rdmp.UI.PieCharts
                 int countPopulated = 0;
                 int countNotPopulated = 0;
                     
-                foreach (ExtractionInformation information in allExtractionInformation)
-                    if (string.IsNullOrWhiteSpace(information.CatalogueItem.Description))
+                foreach (CatalogueItem ci in catalogueItems)
+                    if (string.IsNullOrWhiteSpace(ci.Description))
                         countNotPopulated++;
                     else
                         countPopulated++;
@@ -110,7 +123,6 @@ namespace Rdmp.UI.PieCharts
                 DataTable dt = new DataTable();
                 dt.Columns.Add("Count");
                 dt.Columns.Add("State");
-
 
                 dt.Rows.Add(new object[] { countNotPopulated, "Missing (" + countNotPopulated + ")" });
                 dt.Rows.Add(new object[] { countPopulated, "Populated (" + countPopulated + ")" });
@@ -129,9 +141,20 @@ namespace Rdmp.UI.PieCharts
             }
         }
 
-        public bool ShouldHaveDescription(Catalogue c)
+        private CatalogueItem[] GetCatalogueItems()
         {
-            return !c.IsColdStorageDataset && !c.IsInternalDataset && !c.IsDeprecated && !c.IsProjectSpecific(Activator.RepositoryLocator.DataExportRepository);
+            if (!_collection.IsSingleCatalogueMode)
+            {
+                var catalogues = Activator.CoreChildProvider.AllCatalogues.Where(c => _collection.Include(c, Activator.RepositoryLocator.DataExportRepository)).ToArray();
+
+                //if there are some
+                if (catalogues.Any())
+                    return catalogues.SelectMany(c => c.CatalogueItems).Where(ci => _collection.Include(ci)).ToArray();//get the extractable columns
+                else
+                    return new CatalogueItem[0];//there weren't any so Catalogues so wont be any ExtractionInformationsEither
+            }
+            else
+                return  _collection.GetSingleCatalogueModeCatalogue().CatalogueItems;
         }
 
         public void RefreshBus_RefreshObject(object sender, RefreshObjectEventArgs e)
@@ -140,12 +163,17 @@ namespace Rdmp.UI.PieCharts
         }
 
         bool _bLoading;
+        private bool firstTime = true;
+
         public void SetCollection(IActivateItems activator, IPersistableObjectCollection collection)
         {
             _bLoading = true;
             SetItemActivator(activator);
 
             _collection = (GoodBadCataloguePieChartObjectCollection)collection;
+
+            if(firstTime)
+                SetupFlags();
 
             btnAllCatalogues.Checked = !_collection.IsSingleCatalogueMode;
             btnSingleCatalogue.Checked = _collection.IsSingleCatalogueMode;
@@ -157,6 +185,9 @@ namespace Rdmp.UI.PieCharts
             CommonFunctionality.Add(btnSingleCatalogue);
             CommonFunctionality.Add(btnShowLabels);
             CommonFunctionality.Add(btnRefresh);
+
+            foreach(var mi in _flagOptions)
+                CommonFunctionality.AddToMenu(mi);
 
             GenerateChart();
             _bLoading = false;
@@ -248,24 +279,8 @@ namespace Rdmp.UI.PieCharts
 
         private void btnViewDataTable_Click(object sender, EventArgs e)
         {
-            DataTable dt = new DataTable();
-            dt.Columns.Add("Catalogue");
-            dt.Columns.Add("Count Missing Descriptions");
-            dt.Columns.Add("Missing List");
-            
-            foreach (IGrouping<Catalogue, ExtractionInformation> g in Activator.CoreChildProvider.AllExtractionInformations.GroupBy(ei=>ei.CatalogueItem.Catalogue))
-            {
-                if (!ShouldHaveDescription(g.Key))
-                    continue;
-                
-                var missing = g.Where(ei => string.IsNullOrWhiteSpace(ei.CatalogueItem.Description)).ToArray();
-                dt.Rows.Add(g.Key.Name, missing.Count(), string.Join(",",missing.Select(m=>m.CatalogueItem.Name)));
-            }
-
-            DataTableViewerUI dtv = new DataTableViewerUI(dt,"Catalogue Items Missing Descriptions");
-
-            var form = new SingleControlForm(dtv, true);
-            form.Show();
+            var cmd = new ExecuteCommandShow(Activator, GetCatalogueItems(),1);
+            cmd.Execute();
         }
     }
 }